@@ -1,6 +1,8 @@
 <?php
 namespace App\Http\Routes\V1;
 
+use App\Http\Controllers\V1\Client\AppController;
+use App\Http\Controllers\V1\Client\ClientController;
 use Illuminate\Contracts\Routing\Registrar;
 
 class ClientRoute
@@ -13,11 +15,11 @@ class ClientRoute
         ], function ($router) {
             // Client
             if (empty(config('v2board.subscribe_path'))) {
-                $router->get('/subscribe', 'V1\\Client\\ClientController@subscribe');
+                $router->get('/subscribe', [ClientController::class, 'subscribe']);
             }
             // App
-            $router->get('/app/getConfig', 'V1\\Client\\AppController@getConfig');
-            $router->get('/app/getVersion', 'V1\\Client\\AppController@getVersion');
+            $router->get('/app/getConfig', [AppController::class, 'getConfig']);
+            $router->get('/app/getVersion', [AppController::class, 'getVersion']);
         });
     }
 }
