@@ -2,6 +2,7 @@
 namespace App\Http\Routes\V1;
 
 use Illuminate\Contracts\Routing\Registrar;
+use Illuminate\Support\Facades\App;
 
 class ServerRoute
 {
@@ -11,8 +12,9 @@ class ServerRoute
             'prefix' => 'server'
         ], function ($router) {
             $router->any('/{class}/{action}', function($class, $action) {
-                $ctrl = \App::make("\\App\\Http\\Controllers\\V1\\Server\\" . ucfirst($class) . "Controller");
-                return \App::call([$ctrl, $action]);
+                $controllerClass = "\\App\\Http\\Controllers\\V1\\Server\\" . ucfirst($class) . "Controller";
+                $controller = App::make($controllerClass);
+                return App::call([$controller, $action]);
             });
         });
     }
