@@ -1,6 +1,7 @@
 <?php
 namespace App\Http\Routes\V2;
 
+use App\Http\Controllers\V2\Admin\StatController;
 use Illuminate\Contracts\Routing\Registrar;
 
 class AdminRoute
@@ -12,9 +13,9 @@ class AdminRoute
             'middleware' => ['admin', 'log'],
         ], function ($router) {
             // Stat
-            $router->get ('/stat/override', 'V2\\Admin\\StatController@override');
-            $router->get ('/stat/record', 'V2\\Admin\\StatController@record');
-            $router->get ('/stat/ranking', 'V2\\Admin\\StatController@ranking');
+            $router->get ('/stat/override', [StatController::class, 'override']);
+            $router->get ('/stat/record', [StatController::class, 'record']);
+            $router->get ('/stat/ranking', [StatController::class, 'ranking']);
         });
     }
 }
