@@ -11,6 +11,15 @@ class User extends Model
     protected $guarded = ['id'];
     protected $casts = [
         'created_at' => 'timestamp',
-        'updated_at' => 'timestamp'
+        'updated_at' => 'timestamp',
+        'banned' => 'boolean',
+        'expired_at' => 'timestamp',
+        'transfer_enable' => 'integer',
+        'device_limit' => 'integer',
+        'plan_id' => 'integer',
+        'group_id' => 'integer',
+        'speed_limit' => 'integer',
+        'invite_user_id' => 'integer',
+        'last_login_at' => 'timestamp'
     ];
 }
