@@ -2,6 +2,7 @@
 
 use App\Services\ThemeService;
 use Illuminate\Http\Request;
+use App\Http\Controllers\V1\Client\ClientController;
 
 /*
 |--------------------------------------------------------------------------
@@ -52,5 +53,5 @@ Route::get('/' . config('v2board.secure_path', config('v2board.frontend_admin_pa
 });
 
 if (!empty(config('v2board.subscribe_path'))) {
-    Route::get(config('v2board.subscribe_path'), 'V1\\Client\\ClientController@subscribe')->middleware('client');
+    Route::get(config('v2board.subscribe_path'), [ClientController::class, 'subscribe'])->middleware('client');
 }
